refactoring ready
