+++ .gitignore
@@ -56,6 +56,7 @@ target/
 
 # Jupyter Notebook
 .ipynb_checkpoints
+/*.ipynb
 
 # pyenv
 .python-version
