+++ run/run_FiHy_2018.py
@@ -26,4 +26,4 @@ params = {
 
 resultfile, _ = driver(parameters=params,
                        create_ncf=True,
-                       result_file='FiHy_2018_pa_fix.nc')
+                       result_file='FiHy_2018_photo_fast.nc')
